Remove unnecessary stuff
This change created a single spec failure only caused the first time
specs are run when the test gems installed in `tmp/gems` have not yet
being installed.

The problem was that the code branch that installs the gems uses the
`FileUtils` constant. Accessing that constant makes [this autoload]
trigger the custom rubygems require behavior which will end up setting
gem_home to its current value (when calling `to_spec` on the selected
default gem). At that point, even if the `Gem.clear_paths` call has already
happened, but the old value for `Gem.home` is restored and memoized.
Thus updating `ENV["HOME"]` will have no effect in the path to the
configuration file rubygems uses, and the spec checking that invalid
YAML in the configuration file gives a friendly error will fail, because
the fail is not even loaded.

This could be fixed by removing the `FileUtils` autoload, or stop
memoizing `Gem.home`, but the simplest fix seemed to require `fileutils`
early, so that `Gem.home` is not memoized again after calling
`Gem.clear_paths`.