Tweak RSpec task
If we run RSpec directly, activating the gem suffices. In this case,
RSpec is run via a rake subprocess, we need to make sure ruby is run
with the correct gem path to make sure RSpec dependencies are activated.
Because we just removed the activation from the spec helper file.