Lighther "prefer gems.rb" deprecation
The previous logic is unclear to me. It seemed to try to detect only
"multiple gemfile situations", but it was doing it incorrectly, I think.

The new message is printed _only_ if both gems.rb and Gemfile are
detected in the same project. And recommends sticking with gems.rb. But
we are not yet deprecating "Gemfile" other than that.